when i press spacebar a div is displayed
