+++ src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { HomepageComponent } from './homepage/homepage.component';
-import { GameComponent } from './game/game.component';
-import {OptionsComponent} from './options/options.component'
+import { HomepageComponent } from './pages/homepage/homepage.component';
+import { GameComponent } from './pages/game/game.component';
+import {OptionsComponent} from './pages/options/options.component'
 
 
 const routes: Routes = [
