--- src/app/game/game.component.html
@@ -1,1 +1,0 @@
-<p>game works!</p>
