--- src/app/game/game.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-game',
-  templateUrl: './game.component.html',
-  styleUrls: ['./game.component.css']
-})
-export class GameComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
