+++ src/app/pages/game/game.component.css
@@ -1,0 +1,7 @@
+
+.ammo {
+    height: 10px;
+    width: 4px;
+    background-color: red;
+    position: absolute;
+}
