+++ src/app/pages/game/game.component.html
@@ -1,0 +1,1 @@
+<div *ngIf="shoot" class='ammo'></div>
