+++ src/app/pages/game/game.component.ts
@@ -1,0 +1,24 @@
+import { Component, OnInit, HostListener } from '@angular/core';
+
+@Component({
+  selector: 'app-game',
+  templateUrl: './game.component.html',
+  styleUrls: ['./game.component.css']
+})
+export class GameComponent implements OnInit {
+
+  shoot : boolean = false;
+  @HostListener('document:keydown', ['$event'])
+  onKeydownHandler(event: KeyboardEvent) {
+    
+    if (event.code === 'Space') {
+      this.shoot = true;
+    }
+  }
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
