+++ src/app/shared/ammo.spec.ts
@@ -1,0 +1,7 @@
+import { Ammo } from './ammo';
+
+describe('Ammo', () => {
+  it('should create an instance', () => {
+    expect(new Ammo()).toBeTruthy();
+  });
+});
