+++ src/app/shared/ammo.ts
@@ -1,0 +1,6 @@
+export class Ammo {
+    element: String;
+    posX: Number;
+    posY: Number;
+    color: String;
+}
