+++ src/app/shared/ship.spec.ts
@@ -1,0 +1,7 @@
+import { Ship } from './ship';
+
+describe('Ship', () => {
+  it('should create an instance', () => {
+    expect(new Ship()).toBeTruthy();
+  });
+});
