+++ src/app/shared/ship.ts
@@ -1,0 +1,7 @@
+export class Ship {
+    id: Number;
+    url: String;
+    posX: Number;
+    posY: Number;
+    HP: number;
+}
